+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.9, 2024-01-21
 ---------------
 - ignore swapping blank or otherwise identical slots in inventory and army
+- tweak status message durations
 
 1.8, 2024-01-11
 ---------------